Converts CBS output to signal format

Converts output of run_cbs into the format of the signal files (normal and tumor) used elsewhere in this package. Fills female Y chromosome values with NAs.

Usage

convert_cbs(cbs_out, sif_df, tsig_df)

Arguments

cbs_out

The output from run_cbs (i.e. list of outputs from segment).

sif_df

Tibble or filepath to a text file containing sample metadata

tsig_df

Tibble or filepath to a text file containing the tumor signal matrix

Value

Tumor signal matrix in the same format as the input tumor signal matrix.