Plot Chromosome X and Y signals before and after linear transformation

Plot Chromosome X and Y signals before and after linear transformation

Usage

plot_x_and_y_signal_transformation(
  sif,
  n.df,
  t.df,
  n.df.transformed,
  t.df.transformed,
  output_dir = NULL
)

Arguments

sif

Tibble or data.frame containing sample metadata.

n.df

Tibble or data.frame containing the normal signal matrix. Rownames should be locus names in the format "{chr}:{start}-{end}" where {chr} is 1-22, X, or Y and where {start} and {end} are genomic coordinates

t.df

Tibble or data.frame containing the tumor signal matrix. Rownames should match n.df.

n.df.transformed

Tibble or data.frame containing the normal signal matrix after linear transformation. Output of linear_transformation.

t.df.transformed

Tibble or data.frame containing the tumor signal matrix after linear transformation. Output of linear_transformation.

output_dir

Directory to save the plots. If NULL, the plots will be printed to the screen.

Value

Nothing. Generates and saves or prints plots.