Read in a data frame or filepath with a locus column and arrange it in order of increasing genomic position (1-22, X, Y; within each chromosome, by start position).

Read in a data frame or filepath with a locus column and arrange it in order of increasing genomic position (1-22, X, Y; within each chromosome, by start position).

Usage

read_and_format_input(input, locus_to_rownames = FALSE)

Arguments

input

A tibble or a filepath to a text file containing the data. Must have a locus column in the format "{chr}:{start}-{end}" where {chr} is 1-22, X, or Y and where {start} and {end} are genomic coordinates.

locus_to_rownames

If TRUE, the locus column will be converted to rownames.

Value

Input data frame arranged by genomic position