Run the PseudoTangentXY algorithm

Pseudotangent also isolates the tumor signals using the Tangent algorithm, but it should be used when the set of normals are particularly non-representative.

Usage

run_pseudotangent(
  sif_df,
  nsig_df,
  tsig_df,
  n_latent_init,
  num_partitions,
  n_latent_part,
  cbs_a = 0.005,
  cbs_mw = 3,
  cbs_ncores = 1,
  partition_seed = 37,
  make_plots = FALSE,
  output_dir = NULL
)

Arguments

sif_df

Tibble or filepath to a text file containing sample metadata

nsig_df

Tibble or filepath to a text file containing the normal signal matrix

tsig_df

Tibble or filepath to a text file containing the tumor signal matrix

n_latent_init

The number of latent factors to reconstruct the initial normal subspace

num_partitions

The number of partitions to create in the pseudotangent pipeline. This must be less than the number of tumors

n_latent_part

The number of latent factors for each of the partition runs. This should not exceed the minimum number of normal/pseudonormal samples across all partitions.

cbs_a

The alpha parameter for the function run_cbs

cbs_mw

The minimum width parameter for the function run_cbs

cbs_ncores

The number of cores to use for the function run_cbs

partition_seed

The seed to set for reproducibility for the random partitioning

make_plots

If TRUE, generate plots of latent factor importance and effects of linear transformation

output_dir

Directory to save the plots. If NULL, the plots will be printed to the screen.

Value

A normalized tumor signal matrix