Run the TangentXY algorithm

Isolate tumor signal by removing systematic noise using a set of normal samples.

Usage

run_tangent(
  sif_df,
  nsig_df,
  tsig_df,
  n_latent,
  make_plots = TRUE,
  output_dir = NULL
)

Arguments

sif_df

Tibble or filepath to a text file containing sample metadata

nsig_df

Tibble or filepath to a text file containing the normal signal matrix

tsig_df

Tibble or filepath to a text file containing the tumor signal matrix

n_latent

Number of latent factors to reconstruct normal subspace

make_plots

If TRUE, generate plots of latent factor importance and effects of linear transformation

output_dir

Directory to save the plots. If NULL, the plots will be printed to the screen.

Value

A normalized tumor signal matrix

Examples

res <- run_tangent(example_sif, example_nsig_df, example_tsig_df, 5)
#> Error: object 'example_sif' not found
head(res)
#> Error: object 'res' not found